api/requirements.txt
Flask == 1.1.2
mysql-connector-python == 8.0.27
sqlalchemy >= 1.4.0, <= 1.5.0
pymysql == 1.0.2
pyyaml == 6.0
cryptography == 3.3.2
flask-cors